refactor: Overhaul project structure and add core features
This extensive commit refactors the project by removing the `utils` package and integrating its functionality into more appropriate layers. It also introduces significant new features, enhances existing ones, and improves the overall architecture.

### Major Architectural Changes
- **Deleted `utils` package**: Removed `Constants.kt`, `DateTimeFormatter.kt`, `Extensions.kt`, `ImageManager.kt`, `PasswordHasher.kt`, and `SessionManager.kt`. Their functionalities have been moved to more suitable locations (e.g., `SessionManager` logic is now in `data.local.PreferencesManager`).
- **Introduced `DatabaseSeeder`**: Added a `DatabaseSeeder` to populate the database with initial data on application startup, including a default admin user.
- **Theming and Resources**:
    - Completely overhauled `colors.xml` and `themes.xml` to implement a Material 3 design system.
    - Added a comprehensive set of strings to `strings.xml` for better localization and consistency.

### New Features & Enhancements
- **Test Submission Logic**:
    - `SubmitTestUseCase` now calculates and saves a detailed `TestResult` upon test completion, including score, correct/wrong answers, and time spent.
- **Metrics Calculation**:
    - `ResultRepositoryImpl` now implements logic to calculate performance metrics by subject and by user.
- **Question & Answer Loading**:
    - Repositories (`TestRepositoryImpl`, `QuestionRepositoryImpl`) now correctly load associated answers when fetching questions, providing complete data to the domain layer.
- **MPAndroidChart Integration**: Added the MPAndroidChart library for future data visualization.

### Refinements and Fixes
- **Use Cases**:
    - Added detailed logging and validation to `CreateTestUseCase`.
    - Improved parameter handling in `AssignTestToUsersUseCase` and `StartTestAttemptUseCase`.
- **Splash & Login**:
    - Reworked `SplashActivity` to use a `SessionManager` for checking authentication status.
    - Updated `LoginActivity` and `MainActivity` to use ViewBinding and align with the new navigation structure.
- **Repositories and DAOs**:
    - Added `getAllQuestions` to `QuestionRepository`.
    - Updated DAOs to use `Long` for IDs and added new queries for fetching metrics and counts.
- **UI Layouts**:
    - Updated numerous XML layouts (`fragment_test_list`, `fragment_test_execution`, `fragment_test_result_detail`, etc.) to support new features, improve UI, and fix inconsistencies.
    - Switched from `Fragment` transactions to the Navigation Component in `MainActivity`.

### Build & Configuration
- **Dependencies**: Added dependencies for MPAndroidChart, Kotlinx Serialization, and updated the Hilt version.
- **Gradle**: Enabled ViewBinding and added the JitPack repository.
- **Android Manifest**: Configured `ExamApplication` and updated the activity launch order.

+++ app/src/main/java/com/octal/examly/data/local/dao/AnswerDao.kt
@@ -12,7 +12,7 @@ interface AnswerDao {
     suspend fun insertAll(answers: List<AnswerEntity>): List<Long>
 
     @Query("SELECT * FROM answers WHERE questionId = :questionId")
-    suspend fun getByQuestionId(questionId: Int): List<AnswerEntity>
+    suspend fun getByQuestionId(questionId: Long): List<AnswerEntity>
 
     @Query("SELECT * FROM answers WHERE questionId = :questionId AND isCorrect = 1")
     suspend fun getCorrectAnswersByQuestionId(questionId: Int): List<AnswerEntity>
